Renamed users.tradeRequestsLeft to users.requests_left. Having difficulty completing deleteAllOffersByUserId
