+++ gameswap/database/seedData/users.json
@@ -3,84 +3,84 @@
     "email": "saputra_albert@hotmail.com",
     "username": "albert",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "liezelanne9@gmail.com",
     "username": "liezel",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "hangtimee@gmail.com",
     "username": "jesse",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "nathanaelmullins@gmail.com",
     "username": "nate",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "jjjardine@gmail.com",
     "username": "whitejeff",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "calvin.yao21@gmail.com",
     "username": "calvin",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "charles.spinosa@gmail.com",
     "username": "charlie",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "adam_reback@yahoo.com",
     "username": "adam",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "ohannieho@gmail.com",
     "username": "annie",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "akim2214@gmail.com",
     "username": "tony",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "danielsu@protonmail.ch",
     "username": "daniel",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "fred_david239@yahoo.com",
     "username": "fred",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "tracyyu93@g.ucla.edu",
     "username": "tracy",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   },
   {
     "email": "gujralu@yahoo.com",
     "username": "uttej",
     "password": "",
-    "tradeRequestsLeft": 5
+    "requests_left": 5
   }
 ]
