+++ gameswap/server/controller.js
@@ -1,36 +1,11 @@
 const { Users, Games, Platforms, Offers} = require('../database/model');
+const { consoleIds, helpers } = require('./dbHelpers');
 
-let consoleIds = {
-  4: "Nintendo 64",
-  7: "PlayStation",
-  8: "PlayStation 2",
-  9: "PlayStation 3",
-  11: "Xbox",
-  12: "Xbox 360",
-  18: "Nintendo Entertainment System (NES)",
-  19: "Super Nintendo Entertainment System (SNES)",
-  20: "Nintendo DS",
-  21: "Nintendo GameCube",
-  36: "Xbox Live Arcade",
-  37: "Nintendo 3DS",
-  38: "PlayStation Portable",
-  45: "PlayStation Network",
-  46: "PlayStation Vita",
-  47: "Virtual Console (Nintendo)",
-  48: "PlayStation 4",
-  49: "Xbox One",
-  130: "Nintendo Switch",
-  131: "Nintendo PlayStation",
-  137: "New Nintendo 3DS",
-  159: "Nintendo DSi",
-  160: "Nintendo eShop",
-  165: "PlayStation VR",
-  167: "PlayStation 5"
-  };
+
 
 const controller = {
   addNewUser: (req, res) => {
-    req.body.tradeRequestsLeft = 5;
+    req.body.requests_left = 5;
     Users
     .create(req.body)
     .then(data => res.status(200).send(data))
@@ -55,9 +30,32 @@ const controller = {
     .then((offers) => res.send(offers))
     .catch(err => res.status(500).send(`Error finding all Offers by ${user_id}: `, err))
   },
-  deleteAllOffersByUserId: (req, res) => {
-
-  },
+  // deleteAllOffersByUserId: (req, res) => {
+  //   let { user_id } = req.params;
+  //   Offers
+  //   .findAll({ where: { user_id } })
+  //   .then((arrayData) => {
+  //     let deletePromises = [];
+  //     arrayData.forEach((offer) => {
+  //       let { offer_id, game_id } = offer;
+  //       deletePromises.push(helpers.deleteOffer(offer_id, game_id))
+  //     })
+  //     console.log(deletePromises)
+  //     Promise.all(deletePromises)
+  //     .then((gameIds) => {
+  //       console.log(gameIds)
+  //       let updatePromises = [];
+  //       gameIds.forEach((game_id) => {
+  //         updatePromises.push(helpers.updateNumberOfOffers(game_id))
+  //       })
+  //       console.log(gameIds)
+  //       Promise.all(updatePromises)
+  //       .then(values => res.send(values))
+  //       .catch('DIDNT WORK :(')
+  //     })
+  //     .catch(err => res.send(err))
+  //   })
+  // },
 
   findAllOffersSortedByNew: (req, res) => {
     Offers
@@ -96,6 +94,13 @@ const controller = {
 
     })
     .catch(err => res.status(500).send(`Error deleting offer id #${offer_id}: `, err))
+  },
+  updateGameOffers: (req, res, offers) => {
+  let game_id = req.body.id;
+  Games
+    .update({ offers }, { where: { game_id } })
+    .then(() => controller.findGame(req, res)) // this is to confirm that the update was made
+    .catch(err => res.status(500).send('Error updating offer number in Games table: ', err))
   },
 
   findAllGames: (req, res) => {
@@ -138,13 +143,6 @@ const controller = {
     .catch(err => res.status(500).send('Error finding in Games table: ', err))
   },
 
-  updateGameOffers: (req, res, offers) => {
-    let game_id = req.body.id;
-    Games
-      .update({ offers }, { where: { game_id } })
-      .then(() => controller.findGame(req, res)) // this is to confirm that the update was made
-      .catch(err => res.status(500).send('Error updating offer number in Games table: ', err))
-  },
   addNewGame: (req, res) => {
     req.body.cover = req.body.cover.url.substring(2); // trims cover url
     req.body.platforms = req.body.platforms.filter(i => consoleIds[i]); // sets platforms property
